requirements.txt
Flask
tensorflow
opencv-python
Pillow
kaggle
scipy
flask-cors